requirements.txt
PyYAML~=6.0.3
sentence-transformers~=5.4.1
python-docx~=1.2.0
qdrant-client~=1.17.1
fastapi~=0.110.0
uvicorn~=0.29.0

pydantic~=2.13.2
requests~=2.33.1